+++ packages/plugins/plugin-pos/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @porulle/plugin-pos
 
+## 0.10.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @porulle/core@0.10.8
+  - @porulle/db@0.10.8
+
 ## 0.10.6
 
 ### Patch Changes