+++ packages/plugins/plugin-reviews/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @porulle/plugin-reviews
 
+## 0.10.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @porulle/core@0.10.8
+
 ## 0.10.6
 
 ### Patch Changes